// src/services/model/getSegmentSlug200DataItemCategoriesItem.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * TrinityKitWP API
 * API para criar posts e fazer upload de imagens no WordPress
 * OpenAPI spec version: 1.0.0
 */
import type { GetSegmentSlug200DataItemCategoriesItemProductsItem } from './getSegmentSlug200DataItemCategoriesItemProductsItem';

export type GetSegmentSlug200DataItemCategoriesItem = {
  /** Nome da subcategoria */
  name?: string;
  /** Slug da subcategoria */
  slug?: string;
  products?: GetSegmentSlug200DataItemCategoriesItemProductsItem[];
};