// src/services/model/getSegmentSlug200DataItemCategoriesItemProductsItem.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * TrinityKitWP API
 * API para criar posts e fazer upload de imagens no WordPress
 * OpenAPI spec version: 1.0.0
 */

export type GetSegmentSlug200DataItemCategoriesItemProductsItem = {
  /** ID do produto */
  id?: number;
  /** Título do produto */
  title?: string;
  /** Número CAS do produto */
  cas_number?: string;
};